requirements.txt
certifi==2020.12.5
chardet==4.0.0
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
h11==0.12.0
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
motor==2.3.1
pydantic==1.8.1
pymongo==3.11.3
python-dotenv==0.17.0
python-http-client==3.3.2
python-multipart==0.0.5
requests==2.25.1
sendgrid==6.6.0
six==1.15.0
starkbank-ecdsa==1.1.0
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
wincertstore==0.2
